+++ package.json
@@ -9,10 +9,17 @@
         "lint:js": "eslint ./src/**/*.js",
         "lint:css": "stylelint ./src/**/*.css",
         "lint:ts": "tsc --noEmit",
-        "lint": "npm run lint:ts && npm run lint:css && npm run jest",
+        "lint": "npm run lint:ts && npm run lint:css && npm run test",
         "prepare": "husky install",
         "build": "set NODE_ENV=production&&npx webpack",
         "start": "webpack-dev-server --open"
+    },
+    "jest": {
+        "transform": {
+          ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+        },
+        "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
+        "moduleFileExtensions": ["ts", "tsx", "js"]
     },
     "repository": {
         "type": "git",
@@ -28,6 +35,7 @@
         "http-server": "^14.1.1"
     },
     "devDependencies": {
+        "@types/jest": "^28.1.6",
         "copy-webpack-plugin": "^11.0.0",
         "css-loader": "^6.7.1",
         "css-minimizer-webpack-plugin": "^4.0.0",
@@ -36,7 +44,7 @@
         "eslint-plugin-prettier": "^4.0.0",
         "html-webpack-plugin": "^5.5.0",
         "husky": "^8.0.1",
-        "jest": "^28.1.2",
+        "jest": "^28.1.3",
         "mini-css-extract-plugin": "^2.6.1",
         "prettier": "^2.7.0",
         "style-loader": "^3.3.1",
@@ -44,6 +52,7 @@
         "stylelint-config-prettier": "^9.0.3",
         "stylelint-config-standard": "^26.0.0",
         "stylelint-prettier": "^2.0.0",
+        "ts-jest": "^28.0.7",
         "ts-loader": "^9.3.1",
         "typescript": "^4.7.4",
         "webpack": "^5.73.0",
